seperated the endpoints
